feat: Voice management extracted from GameCoordinator
